+++ packages/@runlightyear/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @runlightyear/cli
 
+## 0.3.0
+
+### Minor Changes
+
+- Updates to work with new production version of lightyear
+
 ## 0.2.0
 
 ### Minor Changes